Move charts to separate file
