+++ app/javascript/lib/chart-financial.js
