+++ app/javascript/packs/application.js
@@ -7,10 +7,6 @@ require("@rails/ujs").start()
 require("turbolinks").start()
 require("@rails/activestorage").start()
 require("channels")
-import Chart from 'chart.js/auto';
-import { DateTime } from "luxon";
-import 'chartjs-adapter-luxon';
-import './chart-financial';
 
 // Uncomment to copy all static images under ../images to the output folder and reference
 // them with the image_pack_tag helper in views (e.g <%= image_pack_tag 'rails.png' %>)
@@ -45,38 +41,3 @@ document.addEventListener("turbolinks:load", function() {
         })
     })
 });
-
-document.addEventListener('turbolinks:load', () => {
-    var ctx = document.getElementById('myChart').getContext('2d');
-    var ctx2 = document.getElementById('myChart2').getContext('2d');
-    var myChart = new Chart(ctx, {
-    type: 'line',
-    data: {
-      labels: JSON.parse(ctx.canvas.dataset.labels),
-      datasets: [{
-        label: 'SOL Token Price',
-        data: JSON.parse(ctx.canvas.dataset.data),
-        fill: false,
-        borderColor: 'rgb(75, 192, 192)',
-        tension: 0.1
-      }]
-    },
-  });
-
-  const data2 = JSON.parse(ctx2.canvas.dataset.data)
-  // Converts string timestamp to Int for luxon
-  data2.forEach(el => el['x'] = parseInt(el['x']));
-
-  var myChart2 = new Chart(ctx2, {
-    type: 'candlestick',
-    data: {
-      datasets: [{
-        label: 'SOL Token Price',
-        data: data2,
-        fill: false,
-        borderColor: 'rgb(75, 192, 192)',
-        tension: 0.1
-      }]
-    },
-  });
-})
