+++ app/views/sol_prices/index.html.erb
@@ -34,3 +34,4 @@
     </div>
   </div>
 </div>
+<%= javascript_pack_tag 'sol_prices_charts' %>
